商品参与活动模块

+++ src/main/java/com/kingcall/seckill/entity/PromoAction.java
@@ -2,6 +2,7 @@ package com.kingcall.seckill.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -11,5 +12,5 @@ public class PromoAction {
     private Date startTime;
     private Date endTime;
     private Integer itemId;
-    private Long promoPrice;
+    private BigDecimal promoPrice;
 }
